[SuperResolution] Fix benchmark kernel mismatch, decouple model specs, add HTML eval harness
- src/benchmark.py: generate synthetic LR with PIL's bicubic filter
  instead of cv2's -- the pretrained checkpoints from #27 were trained
  on PIL-bicubic-degraded LR, and cv2's kernel made them score worse
  than bilinear (kernel-mismatch artifact, not a model defect).
- --models entries are now "label:arch:scale:weights_path", decoupling
  label/arch/scale so one report can compare different archs at the
  same scale or the same arch across scales; results grouped by
  (scale, method) since PSNR/SSIM are only comparable within a scale.
- src/train.py: psnr() now returns a plain float (was leaking
  np.float64/np.bool_ through comparisons, breaking JSON serialization
  downstream).
- test/eval_medicine_package.py: HTML+JSON eval harness for the
  medicine_package fixture, following the same convention as
  research/OCR and research/ObjectDetection's eval scripts (versioned
  run IDs, base64-embedded images, commit-stamped header).

+++ research/SuperResolution/src/train.py
@@ -78,7 +78,7 @@ def psnr(pred: np.ndarray, target: np.ndarray) -> float:
     mse = np.mean((pred.astype(np.float64) - target.astype(np.float64)) ** 2)
     if mse == 0:
         return float("inf")
-    return 20 * np.log10(255.0) - 10 * np.log10(mse)
+    return float(20 * np.log10(255.0) - 10 * np.log10(mse))
 
 
 def train(
